добавлены обработчики ошибок для авторов, также подключено логгирование ошибок для разработчиков

+++ src/services/author.py
@@ -2,6 +2,11 @@ from src.schemas.author import AuthorSchemaAdd, AuthorSchemaDelete, AuthorSchema
 
 from src.utils.uow import AbstractUnitOfWork
 
+from src.utils.exceptions import (
+    ResourceNotFoundException,
+    InvalidIDException,
+)
+
 
 class AuthorService:
     def __init__(self, uow: AbstractUnitOfWork):
@@ -16,16 +21,24 @@ class AuthorService:
     async def get_all(self):
         async with self.uow as uow:
             authors = await uow.authors.get_all()
+            if not authors:
+                raise ResourceNotFoundException("Авторы не найдены.")
             return authors
 
     async def delete_one(self, author_id: AuthorSchemaDelete):
         author_id = author_id.model_dump()["id"]
+        if author_id < 1:
+            raise InvalidIDException(f"Переданное значение id {author_id} недопустимо. Оно должно быть больше 0.")
         async with self.uow as uow:
             author = await uow.authors.delete_one(author_id)
+            if author is None:
+                raise ResourceNotFoundException(f"Автор с id {author_id} не найден.")
             return author
 
     async def update_one(self, author: AuthorSchemaUpdate):
         author_dict = author.model_dump()
+        if author_dict["id"] < 1:
+            raise InvalidIDException(f"Переданное значение id {author_dict['id']} недопустимо. Оно должно быть больше 0.")
         async with self.uow as uow:
             author_id = await uow.authors.update_one(author_dict)
             return author_id
